% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsct_data.R
\docType{data}
\name{hsct}
\alias{hsct}
\title{HSCT data}
\format{A \code{data.frame} contains the following columns:
\describe{
  \item{\code{id}}{patient id}
  \item{\code{Time}}{observation time}
  \item{\code{event}}{event indicator, indicating whether there is a recurrent event}
  \item{\code{status}}{status indicator, indicating whether the terminal event is censored}
  \item{\code{m}}{Number of recurrent event per subject}
  \item{\code{scaleAge}}{scaled age}
  \item{\code{scaleAge}}{positive part of the scaled age, used to remove non-linear effect from age}
  \item{\code{race0}}{1 = white, 0 = otherwise}
  \item{\code{allo}}{1 = allogeneic, 0 = autologous}
  \item{\code{gender}}{1 = male, 0 = female}
  \item{\code{lym}}{1 = lymphomas disease at transplant, 0 = otherwise}
  \item{\code{agvhd}}{This should be a time-dependent covariate; don't use this}
  \item{\code{heme1}}{1 = heme at remission, 0 = otherwise}
  \item{\code{heme2}}{1 = heme at relapse, 0 = otherwise}
  \item{\code{cmv1}}{1 = receiver negative, 0 = otherwise}
  \item{\code{cmv2}}{1 = receiver and donor both negative, 0 = otherwise}
}}
\usage{
data(hsct)
}
\description{
A data frame contains data from HSCT. Only for internal testing use.
Will not share in public.
}
\examples{
data(hsctr)
}
