% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gsm}
\alias{gsm}
\title{Main estimation function}
\usage{
gsm(dat, shp.ind = FALSE, B = 100)
}
\arguments{
\item{dat}{data frame generated from \code{simDat}.}

\item{shp.ind}{a character string indicating either to proceed with shape independence assumption.}

\item{B}{a numerical value specifying number of bootstrap in shape independence testing.}
}
\description{
This function assumes the input data is generated from \code{simDat}.
The suage can be generated to take formula form.
}
