% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{gsm}
\alias{gsm}
\title{Main estimation function}
\usage{
gsm(dat)
}
\arguments{
\item{dat}{data frame generated from \code{simDat}.}
}
\description{
This function assumes the input data is generated from \code{simDat}.
The suage can be generated to take formula form.
}
