% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDat.R
\name{simDat}
\alias{simDat}
\title{Function to generate simulated data}
\usage{
simDat(n, model, frailty = FALSE, type1 = FALSE)
}
\description{
We formulate the proposed rate function as
\deqn{\mu(t|Z) = f(t, \beta_0^\top Z) g(\gamma_0^\top Z), 0\le t \le \tau}.
See Details.
}
\details{
The function \code{simDat} generates simulated recurrent event data from our model's special cases.
The following model are implemented.
\describe{
  \item{\code{type = "M1"}}{generates recurrent event data from the proportional rate model.
    \deqn{\mu(t|Z) = \mu_0(t) e^{\gamma_0^\top Z}.}
     This model implies \eqn{\beta_0 = 0}.}
  \item{\code{type = "M2"}}{generates recurrent event data from the additive model.
    \deqn{\mu(t|Z) = \mu_0(t) + \alpha_0^\top Z.}
     This model implies \eqn{\beta_0 = \gamma_0}.}
  \item{\code{type = "M3"}}{generates recurrent event data from the accelerated rate model.
    \deqn{\mu(t|Z) = \mu_0(t e^{\alpha_0^\top Z}).}
     This model implies \eqn{\beta_0 = \gamma_0}.}
  \item{\code{type = "M4"}}{generates recurrent event data from the two-sample model.
    \deqn{\mu(t|Z) = \mu_0(t e^{\beta_0^\top Z})e^{\gamma_0^\top Z}.}}
}
The data frame is similar to the one used in \code{reReg}.
The data consists of the following columns:
\describe{
  \item{id}{Patient ID}
  \item{t}{Patient's event time; this column consists of both the recurrent event time and the terminal event time}
  \item{m}{The observed number of recurrent events}
  \item{x1, x2}{Baseline covariates}
  \item{event}{recurrent event indicator; 1 = recurrent event, 0 = not a recurrent event}
  \item{status}{censoring indicator; this is only meanful for when event = 0}
}
For all scenarios, we set \eqn{\tau = 10}, \eqn{\beta_0 = (0.6, 0.8)} and \eqn{\gamma_0 = (7/25, 24/25)}.
We also set \eqn{\mu_0(t) = \frac{2}{1 + t}}, then \eqn{m_0(t) = 2\log(1 + t)}.
}
