% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDat.R
\name{simDat}
\alias{simDat}
\title{Function to generate simulated data}
\usage{
simDat(n, model, frailty = FALSE, ...)
}
\arguments{
\item{n}{is the sample size}

\item{model}{is a character string indicating the scenario;
available options are M1, M2, M3, and M4, corresponding to models M1, M2, M3, and M4
described in the manuscript.}

\item{frailty}{is a logical value indicating whether to generate
the data under informative censoring}
}
\description{
The function \code{simDat} is used to generate recurrent event data used in the simulation of the manuscript.
}
\details{
The data frame is similar to the one used in \code{reReg}.
The data consists of the following columns:
\describe{
  \item{id}{Patient ID}
  \item{t}{Patient's event time; this column consists of both the recurrent event time and the terminal event time}
  \item{m}{The observed number of recurrent events}
  \item{x1, x2}{Baseline covariates}
  \item{event}{recurrent event indicator; 1 = recurrent event, 0 = not a recurrent event}
  \item{status}{censoring indicator; this is only meanful for when event = 0}
}
}
